(1) merge allow_inline and fast_execution_in_network_thread so that allow_line is applicable to all tasks (2) remove throttling by execution time estimation and rejection, since its runtime overhead is too high and there is risks where estimation is not precise which leads to lots of false positives; (3) recover throttling mechanism by delaying receiving rpc session's recv ops. Now for edge services, we can use reject, and for internal services, we can use delay. (4) always use io-looper-timer instead of simple-timer-service in fastrun.

+++ src/core/core/rpc_message.cpp
@@ -378,7 +378,8 @@ message_ex* message_ex::copy_and_prepare_send()
         // the message_header is hidden ahead of the buffer, expose it to buffer
         dassert(buffers.size() == 1, "there must be only one buffer for read msg");
         dassert((char*)header + sizeof(message_header) == (char*)buffers[0].data(), "header and content must be contigous");
-        copy->buffers[0] = copy->buffers[0].range(-sizeof(message_header));
+
+        copy->buffers[0] = copy->buffers[0].range(-(int)sizeof(message_header));
 
         // switch the flag
         copy->_is_read = false;
